Make source distributions cross-platform
The manifest file included all .h files by default, but not all .c files. The only .c files included in source distributions were the ones specified as the sources for the C extension, which were platform specific. This made the source distributions work only on the platform they were made on.